laosChatbotAI01 README.md
I have, on the verge of my last "production", created a chatbot that will be able to understand user input and provide appropriate responses based on predefined intents. 
For example, if the user asks if a store is open after hours, we would define intents and then have a predefined answer for those specific keywords. 
We also analyze the sentiment of the user's messages to generate responses that match the emotional tone.